# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API Sprint 5",
        "description": "Documentación de la API",
        "contact": {
            "email": "ivonnecp@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/photos": {
            "get": {
                "tags": [
                    "Fotos"
                ],
                "summary": "Obtener todas las fotos",
                "operationId": "793cf02e220bab257709ded15be493c2",
                "responses": {
                    "200": {
                        "description": "Lista de fotos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "photos": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Fotos"
                ],
                "summary": "Crear una nueva foto",
                "operationId": "e21993f5b41f1c66ee580264d09c4cea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "gallery_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "gallery_id": {
                                        "type": "integer"
                                    },
                                    "location": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto subida satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "photo": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/photos/{photo}": {
            "get": {
                "tags": [
                    "Fotos"
                ],
                "summary": "Obtener una foto específica",
                "operationId": "271d6770e3400d50b59250f531c7f43c",
                "parameters": [
                    {
                        "name": "photo",
                        "in": "path",
                        "description": "ID de la foto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la foto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "photo": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Fotos"
                ],
                "summary": "Actualizar una foto",
                "operationId": "e2d69df94c930ebdde29461819af2834",
                "parameters": [
                    {
                        "name": "photo",
                        "in": "path",
                        "description": "ID de la foto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "gallery_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "gallery_id": {
                                        "type": "integer"
                                    },
                                    "location": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto actualizada satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "photo": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
        "/api/usuarios/{user}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "2f5c62e525afbbad64d3e1d35f4ce595",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "3ab8fe6aac5964611fd4d589e666552d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No puedes eliminarte a ti mismo"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
...
}